// src/types/shared.ts
export type ColumnDataType = ObjectValues<typeof COLUMN_TYPE> | ''

export interface FormatOptions {
  showDate?: boolean
  showTime?: boolean
  locale?: string
  statusSwitch?: any
}

export interface ColumnDataField<T> {
  field: keyof T
  params?: any
  class?: string
  type?: ColumnDataType
  renderer?: (data: any) => ReactNode
  onToggle?: (row: T) => void
  checkCondition?: (val: any, row: T) => boolean
  dateformatOptions?: FormatOptions
  formatOptions?: any
}

export type Column<T> = {
  title: string
  sortField?: keyof T
  dataField: ColumnDataField<T>[]
  class?: string
  sortable?: boolean
  sortDirection?: 'asc' | 'desc'
  hidden?: boolean
}

export interface ViewConfig<T> {
  redirectUrl: (data: T) => string
  icon?: string
  className?: string
}
export interface Action<T = any> {
  label: string
  actionToPerform?: string
  icon?: string
  conditional?: Conditional<T>
  getNavigateUrl?: (data: T) => string
  className?: string
  renderer?: (row: any) => ReactNode
  viewConfig?: ViewConfig<T>
}

export interface Conditional<T> {
  field: keyof T
  condition: '!=' | '?' | 'include' | 'notInclude' | '==='
  conditionValue?: string[]
  actionLabel?: string
}

export interface ITableConfig<T = any> {
  columns: Column<T>[]
  actionsDropDown: (Action<T> | string)[]
  data: T[] | undefined
  total: number | undefined
  isColumnChooser?: boolean
}

export type TableConfig<T> = ITableConfig & {
  columns: Column<T>[]
  data: T[] | []
}